Cuda refactor

+++ src/cubes/marching_cubes_common.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <glm/glm.hpp>
+#include <cstdint>
 
 /*
     Tables and conventions from
@@ -14,14 +15,15 @@ const glm::uvec2 edgeToVertices[12] =
     {0, 4}, {1, 5}, {2, 6}, {3, 7}
 };
 
-const unsigned edgeToVerticesU[12][2] = 
+const uint8_t edgeToVerticesU[12][2] = 
 {
     {0, 1}, {1, 2}, {2, 3}, {3, 0},
     {4, 5}, {5, 6}, {6, 7}, {7, 4},
     {0, 4}, {1, 5}, {2, 6}, {3, 7}
 };
 
-const int edgeTable[256] =
+// 12 bits used
+const uint16_t edgeTable[256] =
 {
     0x0  , 0x109, 0x203, 0x30a, 0x406, 0x50f, 0x605, 0x70c,
     0x80c, 0x905, 0xa0f, 0xb06, 0xc0a, 0xd03, 0xe09, 0xf00,
@@ -57,7 +59,7 @@ const int edgeTable[256] =
     0x70c, 0x605, 0x50f, 0x406, 0x30a, 0x203, 0x109, 0x0   
 };
 
-const int triTable[256][16] =
+const int8_t triTable[256][16] =
 {
     {-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
     {0, 8, 3, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
